# microservices/api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gateway - Microservices E-commerce",
    "description": "API Gateway centralisant tous les microservices de l'application e-commerce avec load balancing et sécurité CORS",
    "version": "1.0.0",
    "contact": {
      "name": "LOG430 - Lab5",
      "email": "support@ecommerce.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Gateway local"
    },
    {
      "url": "http://api-gateway:3000",
      "description": "API Gateway Docker"
    }
  ],
  "paths": {
...
    "/api/produits": {
      "get": {
        "summary": "Récupérer tous les produits",
        "description": "Récupère la liste de tous les produits disponibles",
        "tags": ["Produits"],
        "responses": {
          "200": {
            "description": "Liste des produits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Produit"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer un nouveau produit",
        "description": "Ajoute un nouveau produit au catalogue",
        "tags": ["Produits"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduitInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produit"
                }
              }
            }
          }
        }
      }
    },
    "/api/produits/{id}": {
      "get": {
        "summary": "Récupérer un produit par ID",
        "description": "Récupère les détails d'un produit spécifique",
        "tags": ["Produits"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID du produit"
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du produit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produit"
                }
              }
            }
          },
          "404": {
            "description": "Produit non trouvé"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Produit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID unique du produit"
          },
          "nom": {
            "type": "string",
            "description": "Nom du produit"
          },
          "description": {
            "type": "string",
            "description": "Description du produit"
          },
          "prix": {
            "type": "number",
            "description": "Prix du produit"
          },
          "categorie": {
            "type": "string",
            "description": "Catégorie du produit"
          },
          "codeProduit": {
            "type": "string",
            "description": "Code unique du produit"
          },
          "quantiteStock": {
            "type": "integer",
            "description": "Quantité en stock"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProduitInput": {
        "type": "object",
        "required": ["nom", "description", "prix", "categorie", "codeProduit", "quantiteStock"],
        "properties": {
          "nom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "prix": {
            "type": "number"
          },
          "categorie": {
            "type": "string"
          },
          "codeProduit": {
            "type": "string"
          },
          "quantiteStock": {
            "type": "integer"
          }
        }
      },
...
    }
  },
...
} 